requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
dnspython==1.12.0
nose==1.3.4
coverage==3.7.1
mock==1.0.1
pylint==1.3.1
